chatTensorFlow/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain_openai
langchain_anthropic
langchain_cohere
langchain_chroma
chromadb
langgraph
langgraph-sdk
langgraph-cli[inmem]
langsmith
requests
crawl4ai
beautifulsoup4
aiohttp
python-dotenv
numexpr
flask
pytest
pytest-asyncio

